Rename completion providers to actually include `CompletionProvider` suffix
