+++ src/LanguageServer.Engine/CompletionProviders/ItemMetadataExpressionCompletionProvider.cs
@@ -19,16 +19,16 @@ namespace MSBuildProjectTools.LanguageServer.CompletionProviders
     /// <summary>
     ///     Completion provider for item metadata expressions.
     /// </summary>
-    public class ItemMetadataExpressionCompletion
+    public class ItemMetadataExpressionCompletionProvider
         : CompletionProvider
     {
         /// <summary>
-        ///     Create a new <see cref="ItemMetadataExpressionCompletion"/>.
+        ///     Create a new <see cref="ItemMetadataExpressionCompletionProvider"/>.
         /// </summary>
         /// <param name="logger">
         ///     The application logger.
         /// </param>
-        public ItemMetadataExpressionCompletion(ILogger logger)
+        public ItemMetadataExpressionCompletionProvider(ILogger logger)
             : base(logger)
         {
         }
