+++ src/LanguageServer.Engine/CompletionProviders/PropertyElementCompletionProvider.cs
@@ -3,6 +3,7 @@ using Serilog;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -15,18 +16,18 @@ namespace MSBuildProjectTools.LanguageServer.CompletionProviders
     using Utilities;
 
     /// <summary>
-    ///     Completion provider for the common item elements.
+    ///     Completion provider for the common property elements.
     /// </summary>
-    public class ItemElementCompletion
+    public class PropertyElementCompletionProvider
         : CompletionProvider
     {
         /// <summary>
-        ///     Create a new <see cref="ItemElementCompletion"/>.
+        ///     Create a new <see cref="PropertyElementCompletionProvider"/>.
         /// </summary>
         /// <param name="logger">
         ///     The application logger.
         /// </param>
-        public ItemElementCompletion(ILogger logger)
+        public PropertyElementCompletionProvider(ILogger logger)
             : base(logger)
         {
         }
@@ -34,12 +35,7 @@ namespace MSBuildProjectTools.LanguageServer.CompletionProviders
         /// <summary>
         ///     The provider display name.
         /// </summary>
-        public override string Name => "Item Elements";
-
-        /// <summary>
-        ///     The provider's default priority for completion items.
-        /// </summary>
-        public override int Priority => 1000;
+        public override string Name => "Property Elements";
 
         /// <summary>
         ///     Provide completions for the specified location.
@@ -69,13 +65,13 @@ namespace MSBuildProjectTools.LanguageServer.CompletionProviders
 
             List<CompletionItem> completions = new List<CompletionItem>();
 
-            Log.Verbose("Evaluate completions for {XmlLocation:l} (trigger characters = {TriggerCharacters})", location, triggerCharacters);
+            Log.Verbose("Evaluate completions for {XmlLocation:l}", location);
 
             using (await projectDocument.Lock.ReaderLockAsync(cancellationToken))
             {
-                if (!location.CanCompleteElement(out XSElement replaceElement, parentPath: WellKnownElementPaths.ItemGroup))
+                if (!location.CanCompleteElement(out XSElement replaceElement, parentPath: WellKnownElementPaths.PropertyGroup))
                 {
-                    Log.Verbose("Not offering any completions for {XmlLocation:l} (not a direct child of a 'ItemGroup' element).", location);
+                    Log.Verbose("Not offering any completions for {XmlLocation:l} (not a direct child of a 'PropertyGroup' element).", location);
 
                     return null;
                 }
@@ -119,7 +115,7 @@ namespace MSBuildProjectTools.LanguageServer.CompletionProviders
         }
 
         /// <summary>
-        ///     Get item element completions.
+        ///     Get property element completions.
         /// </summary>
         /// <param name="projectDocument">
         ///     The <see cref="ProjectDocument"/> for which completions will be offered.
@@ -134,53 +130,53 @@ namespace MSBuildProjectTools.LanguageServer.CompletionProviders
         {
             LspModels.Range replaceRangeLsp = replaceRange.ToLsp();
 
-            HashSet<string> offeredItemNames = new HashSet<string>
-            {
-                "PackageReference",
-                "DotNetCliToolReference"
-            };
+            HashSet<string> offeredPropertyNames = new HashSet<string>();
 
             // Well-known (but standard-format) properties.
 
-            foreach (string wellKnownItemName in MSBuildSchemaHelp.WellKnownItemTypes)
+            foreach (string wellKnownPropertyName in MSBuildSchemaHelp.WellKnownPropertyNames)
             {
-                if (!offeredItemNames.Add(wellKnownItemName))
+                if (!offeredPropertyNames.Add(wellKnownPropertyName))
                     continue;
 
-                yield return ItemCompletion(wellKnownItemName, replaceRangeLsp,
-                    description: MSBuildSchemaHelp.ForItemType(wellKnownItemName)
+                var (defaultValue, defaultValues) = MSBuildSchemaHelp.DefaultsForProperty(wellKnownPropertyName);
+
+                yield return PropertyCompletionItem(wellKnownPropertyName, replaceRangeLsp,
+                    description: MSBuildSchemaHelp.ForProperty(wellKnownPropertyName),
+                    defaultValue: defaultValue,
+                    defaultValues: defaultValues
                 );
             }
 
             if (!projectDocument.HasMSBuildProject)
                 yield break; // Without a valid MSBuild project (even a cached one will do), we can't inspect existing MSBuild properties.
 
-            if (!projectDocument.Workspace.Configuration.Language.CompletionsFromProject.Contains(CompletionSource.ItemType))
+            if (!projectDocument.Workspace.Configuration.Language.CompletionsFromProject.Contains(CompletionSource.Property))
                 yield break;
 
-            int otherItemPriority = Priority + 10;
+            int otherPropertyPriority = Priority + 10;
 
-            string[] otherItemNames =
+            string[] otherPropertyNames =
                 projectDocument.MSBuildProject.Properties
-                    .Select(item => item.Name)
-                    .Where(itemName => !itemName.StartsWith("_")) // Ignore private item types.
+                    .Select(property => property.Name)
+                    .Where(propertyName => !propertyName.StartsWith("_")) // Ignore private properties.
                     .ToArray();
-            foreach (string itemName in otherItemNames)
+            foreach (string propertyName in otherPropertyNames)
             {
-                if (!offeredItemNames.Add(itemName))
+                if (!offeredPropertyNames.Add(propertyName))
                     continue;
 
-                yield return ItemCompletion(itemName, replaceRangeLsp, otherItemPriority,
-                    description: $"I don't know anything about the '{itemName}' item type, but it's defined in this project (or a project that it imports); you can override its value by specifying it here."
+                yield return PropertyCompletionItem(propertyName, replaceRangeLsp, otherPropertyPriority,
+                    description: $"I don't know anything about the '{propertyName}' property, but it's defined in this project (or a project that it imports); you can override its value by specifying it here."
                 );
             }
         }
 
         /// <summary>
-        ///     Create a standard <see cref="CompletionItem"/> for the specified MSBuild item.
+        ///     Create a standard <see cref="CompletionItem"/> for the specified MSBuild property.
         /// </summary>
-        /// <param name="itemName">
-        ///     The MSBuild item name.
+        /// <param name="propertyName">
+        ///     The MSBuild property name.
         /// </param>
         /// <param name="replaceRange">
         ///     The range of text that will be replaced by the completion.
@@ -189,27 +185,95 @@ namespace MSBuildProjectTools.LanguageServer.CompletionProviders
         ///     The item sort priority (defaults to <see cref="CompletionProvider.Priority"/>).
         /// </param>
         /// <param name="description">
-        ///     An optional description for the item.
+        ///     An optional description for the property.
+        /// </param>
+        /// <param name="defaultValue">
+        ///     An optional default value for the property.
+        /// </param>
+        /// <param name="defaultValues">
+        ///     An optional list of default values for the property.
+        /// 
+        ///     If specified, then the inserted property's snippet will offer these as a drop-down list.
         /// </param>
         /// <returns>
         ///     The <see cref="CompletionItem"/>.
         /// </returns>
-        CompletionItem ItemCompletion(string itemName, LspModels.Range replaceRange, int? priority = null, string description = null)
+        CompletionItem PropertyCompletionItem(string propertyName, LspModels.Range replaceRange, int? priority = null, string description = null, string defaultValue = null, IReadOnlyList<string> defaultValues = null)
         {
             return new CompletionItem
             {
-                Label = $"<{itemName}>",
-                Detail = "Item",
+                Label = $"<{propertyName}>",
+                Detail = "Property",
                 Documentation = description,
-                Kind = CompletionItemKind.Class,
-                SortText = $"{priority ?? Priority:0000}<{itemName}>",
+                Kind = CompletionItemKind.Property,
+                SortText = $"{priority ?? Priority:0000}<{propertyName}>",
                 TextEdit = new TextEdit
                 {
-                    NewText = $"<{itemName} Include=\"$1\" />$0",
+                    NewText = GetCompletionText(propertyName, defaultValue, defaultValues),
                     Range = replaceRange
                 },
                 InsertTextFormat = InsertTextFormat.Snippet
             };
+        }
+
+        /// <summary>
+        ///     Get the completion text for the specified property and its default value(s), if any.
+        /// </summary>
+        /// <param name="propertyName">
+        ///     The property name.
+        /// </param>
+        /// <param name="defaultValue">
+        ///     The property's default value (if any).
+        /// </param>
+        /// <param name="defaultValues">
+        ///     The property's default values (if any).
+        /// 
+        ///     If specified, then the inserted property's snippet will offer these as a drop-down list.
+        /// </param>
+        /// <returns>
+        ///     The completion text (in standard LSP Snippet format).
+        /// </returns>
+        static string GetCompletionText(string propertyName, string defaultValue, IReadOnlyList<string> defaultValues)
+        {
+            StringBuilder completionText = new StringBuilder();
+            completionText.AppendFormat("<{0}>", propertyName);
+
+            bool haveValue = false;
+            if (defaultValues != null && defaultValues.Count > 0)
+            {
+                haveValue = true;
+
+                completionText.Append("${1|");
+                completionText.Append(
+                    defaultValues[0]
+                );
+                for (int valueIndex = 1; valueIndex < defaultValues.Count; valueIndex++)
+                {
+                    completionText.Append(',');
+                    completionText.Append(
+                        defaultValues[valueIndex]
+                    );
+                }
+                completionText.Append("|}");
+            }
+            else if (!string.IsNullOrWhiteSpace(defaultValue))
+            {
+                haveValue = true;
+
+                completionText.Append("${1:");
+                completionText.Append(defaultValue);
+                completionText.Append('}');
+            }
+            else
+                completionText.Append("$0");
+
+            completionText.AppendFormat("</{0}>", propertyName);
+
+            // If we have a default value / values, then the final cursor position should be after the property element.
+            if (haveValue)
+                completionText.Append("$0");
+
+            return completionText.ToString();
         }
     }
 }
